feat: basic user.managemente implemented
Now user-managent keep a .db file that contains id, name and passwod of all users.
It is not called by gateway because the service not can do nothing that not make necesary a validation.
Only can register and login if is called by other service. I tested with adapted auth-service and work, but the service not manage the token and the auth-service is not adapted
to a external service manage tha dates that i say before.