feat: add four endpoints
One point is a getter that use user Id to expose all friends, other that add a friends, other that delete friends and las one that check if two users are friends.

We use a logic in data basament that one line says one friends is a friend of other id. When we introduce a friends, we introduce two lines where A is friend of B and other one that says B is friend of A.