fix: All profile.ts errors fixed.
No errors for profile now. Fixed the game matches request in order to get everything working. Had to add a function in api.ts to get the matches by id to call the endpoint for better readability in the profile.ts. Also added he /game/ to indicate the rotuer that the petition was for the server-side-pong service.